requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
openai==1.63.2
markdown2
python-docx